+++ src/main/java/backend/wal/auth/application/provider/AuthServiceProvider.java
@@ -1,6 +1,6 @@
 package backend.wal.auth.application.provider;
 
-import backend.wal.auth.application.port.AuthUseCase;
+import backend.wal.auth.application.port.in.AuthUseCase;
 import backend.wal.auth.application.service.AppleAuthService;
 import backend.wal.auth.application.service.KakaoAuthService;
 import backend.wal.user.domain.aggregate.SocialType;
