+++ src/main/java/backend/wal/auth/web/AuthController.java
@@ -1,13 +1,13 @@
 package backend.wal.auth.web;
 
-import backend.wal.auth.application.port.dto.TokenResponse;
-import backend.wal.auth.application.port.IssueTokenUseCase;
-import backend.wal.auth.application.port.AuthUseCase;
+import backend.wal.auth.application.port.out.TokenResponseDto;
+import backend.wal.auth.application.port.in.IssueTokenUseCase;
+import backend.wal.auth.application.port.in.AuthUseCase;
 import backend.wal.auth.application.provider.AuthServiceProvider;
 import backend.wal.auth.web.dto.LoginRequest;
 import backend.wal.support.annotation.Authentication;
 import backend.wal.support.annotation.ExtractValidRefreshToken;
-import backend.wal.utils.HttpHeaderUtils;
+import backend.wal.support.utils.HttpHeaderUtils;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -26,11 +26,11 @@ public class AuthController {
     public ResponseEntity<Void> login(@Valid @RequestBody LoginRequest request) {
         AuthUseCase authUseCase = authServiceProvider.getAuthServiceBy(request.getSocialType());
         Long userId = authUseCase.login(request.toAuthServiceDto());
-        TokenResponse tokenResponse = issueTokenUseCase.issue(userId);
+        TokenResponseDto tokenResponseDto = issueTokenUseCase.issue(userId);
 
         return ResponseEntity.ok()
-                .header(HttpHeaderUtils.AUTHORIZATION, HttpHeaderUtils.withBearerToken(tokenResponse.getAccessToken()))
-                .header(HttpHeaderUtils.REFRESH_TOKEN, tokenResponse.getRefreshToken())
+                .header(HttpHeaderUtils.AUTHORIZATION, HttpHeaderUtils.withBearerToken(tokenResponseDto.getAccessToken()))
+                .header(HttpHeaderUtils.REFRESH_TOKEN, tokenResponseDto.getRefreshToken())
                 .build();
     }
